:recycle: refactor: 프로젝트명 변경
- HabitTracking -> HabitTracker

Issue: #14
