+++ app/src/androidTest/java/com/bibbidi/habittracker/ExampleInstrumentedTest.kt
@@ -1,0 +1,16 @@
+package com.bibbidi.habittracker
+
+import androidx.test.ext.junit.runners.AndroidJUnit4
+import androidx.test.platform.app.InstrumentationRegistry
+import org.junit.Assert.assertEquals
+import org.junit.Test
+import org.junit.runner.RunWith
+
+@RunWith(AndroidJUnit4::class)
+class ExampleInstrumentedTest {
+    @Test
+    fun useAppContext() {
+        val appContext = InstrumentationRegistry.getInstrumentation().targetContext
+        assertEquals("com.bibbidi.HabitTracker", appContext.packageName)
+    }
+}
