+++ app/src/main/java/com/bibbidi/habittracker/HabitTrackerApplication.kt
@@ -1,0 +1,7 @@
+package com.bibbidi.habittracker
+
+import android.app.Application
+import dagger.hilt.android.HiltAndroidApp
+
+@HiltAndroidApp
+class HabitTrackerApplication : Application()
