+++ app/src/main/java/com/bibbidi/habittracker/data/entity/DayOfTheWeek.kt
@@ -1,0 +1,11 @@
+package com.bibbidi.habittracker.data.entity
+
+enum class DayOfTheWeek {
+    SUNDAY,
+    MONDAY,
+    TUESDAY,
+    WEDNESDAY,
+    THURSDAY,
+    FRIDAY,
+    SATURDAY
+}
