+++ app/src/main/java/com/bibbidi/habittracker/data/entity/TimeFrame.kt
@@ -1,0 +1,7 @@
+package com.bibbidi.habittracker.data.entity
+
+enum class TimeFrame {
+    MORNING,
+    AFTERNOON,
+    EVENING
+}
