+++ app/src/main/java/com/bibbidi/habittracker/ui/MainActivity.kt
@@ -1,0 +1,20 @@
+package com.bibbidi.habittracker.ui
+
+import android.os.Bundle
+import androidx.appcompat.app.AppCompatActivity
+import com.bibbidi.habittracker.databinding.ActivityMainBinding
+import dagger.hilt.android.AndroidEntryPoint
+
+@AndroidEntryPoint
+class MainActivity : AppCompatActivity() {
+
+    private lateinit var binding: ActivityMainBinding
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        binding = ActivityMainBinding.inflate(layoutInflater)
+        setContentView(binding.root)
+
+        setSupportActionBar(binding.toolbar)
+    }
+}
