+++ app/src/main/java/com/bibbidi/habittracker/ui/customview/DayOfTheWeek.kt
@@ -1,0 +1,11 @@
+package com.bibbidi.habittracker.ui.customview
+
+enum class DayOfTheWeek {
+    SUN,
+    MON,
+    TUE,
+    WED,
+    THU,
+    FRI,
+    SAT
+}
