+++ app/src/main/java/com/bibbidi/habittracker/ui/home/DateItem.kt
@@ -1,0 +1,10 @@
+package com.bibbidi.habittracker.ui.home
+
+import com.bibbidi.habittracker.ui.customview.DayOfTheWeek
+
+data class DateItem(
+    val checked: Boolean,
+    val dayOfTheMonth: Int,
+    val dayOfTheWeek: DayOfTheWeek,
+    val isToday: Boolean
+)
