--- app/src/main/java/com/bibbidi/habittracking/HabitTrackingApplication.kt
@@ -1,7 +1,0 @@
-package com.bibbidi.habittracking
-
-import android.app.Application
-import dagger.hilt.android.HiltAndroidApp
-
-@HiltAndroidApp
-class HabitTrackingApplication : Application()
