--- app/src/main/java/com/bibbidi/habittracking/data/entity/DayOfTheWeek.kt
@@ -1,11 +1,0 @@
-package com.bibbidi.habittracking.data.entity
-
-enum class DayOfTheWeek {
-    SUNDAY,
-    MONDAY,
-    TUESDAY,
-    WEDNESDAY,
-    THURSDAY,
-    FRIDAY,
-    SATURDAY
-}
