--- app/src/main/java/com/bibbidi/habittracking/data/entity/LongMemoInstance.kt
@@ -1,12 +1,0 @@
-package com.bibbidi.habittracking.data.entity
-
-import androidx.room.Entity
-import androidx.room.PrimaryKey
-
-@Entity
-data class LongMemoInstance(
-    @PrimaryKey(autoGenerate = true)
-    val longMemoInstanceId: Long,
-    val habitId: Int,
-    val content: String
-)
