--- app/src/main/java/com/bibbidi/habittracking/data/entity/TimeFrame.kt
@@ -1,7 +1,0 @@
-package com.bibbidi.habittracking.data.entity
-
-enum class TimeFrame {
-    MORNING,
-    AFTERNOON,
-    EVENING
-}
