--- app/src/main/java/com/bibbidi/habittracking/data/entity/check/CheckHabit.kt
@@ -1,22 +1,0 @@
-package com.bibbidi.habittracking.data.entity.check
-
-import androidx.room.Entity
-import androidx.room.ForeignKey
-import androidx.room.PrimaryKey
-import com.bibbidi.habittracking.data.entity.Habit
-
-@Entity(
-    foreignKeys = [
-        ForeignKey(
-            entity = Habit::class,
-            parentColumns = ["id"],
-            childColumns = ["habitId"],
-            onDelete = ForeignKey.CASCADE
-        )
-    ]
-)
-data class CheckHabit(
-    @PrimaryKey(autoGenerate = true)
-    val id: Long,
-    val habitId: Long
-)
