--- app/src/main/java/com/bibbidi/habittracking/ui/MainActivity.kt
@@ -1,20 +1,0 @@
-package com.bibbidi.habittracking.ui
-
-import android.os.Bundle
-import androidx.appcompat.app.AppCompatActivity
-import com.bibbidi.habittracking.databinding.ActivityMainBinding
-import dagger.hilt.android.AndroidEntryPoint
-
-@AndroidEntryPoint
-class MainActivity : AppCompatActivity() {
-
-    private lateinit var binding: ActivityMainBinding
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        binding = ActivityMainBinding.inflate(layoutInflater)
-        setContentView(binding.root)
-
-        setSupportActionBar(binding.toolbar)
-    }
-}
