--- app/src/main/java/com/bibbidi/habittracking/ui/customview/DayOfTheWeek.kt
@@ -1,11 +1,0 @@
-package com.bibbidi.habittracking.ui.customview
-
-enum class DayOfTheWeek {
-    SUN,
-    MON,
-    TUE,
-    WED,
-    THU,
-    FRI,
-    SAT
-}
