--- app/src/main/java/com/bibbidi/habittracking/ui/home/DateItem.kt
@@ -1,10 +1,0 @@
-package com.bibbidi.habittracking.ui.home
-
-import com.bibbidi.habittracking.ui.customview.DayOfTheWeek
-
-data class DateItem(
-    val checked: Boolean,
-    val dayOfTheMonth: Int,
-    val dayOfTheWeek: DayOfTheWeek,
-    val isToday: Boolean
-)
