+++ app/src/main/res/values-night/themes.xml
@@ -1,6 +1,6 @@
 <resources>
     <!-- Base application theme. -->
-    <style name="Theme.HabitTracking" parent="Theme.MaterialComponents.DayNight.NoActionBar">
+    <style name="Theme.HabitTracker" parent="Theme.MaterialComponents.DayNight.NoActionBar">
         <!-- Primary brand color. -->
         <item name="colorPrimary">@color/pink_04</item>
         <item name="colorPrimaryVariant">@color/pink_03</item>
