+++ app/src/main/res/values/themes.xml
@@ -1,6 +1,6 @@
 <resources>
     <!-- Base application theme. -->
-    <style name="Theme.HabitTracking" parent="Theme.MaterialComponents.Light.NoActionBar">
+    <style name="Theme.HabitTracker" parent="Theme.MaterialComponents.Light.NoActionBar">
         <!-- Primary brand color. -->
         <item name="colorPrimary">@color/pink_03</item>
         <item name="colorPrimaryVariant">@color/pink_02</item>
